# d1/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API REST (LARAVEL 7) IDESA ",
        "description": "API REST construido con Laravel 7 y Documentado con Swagger",
        "contact": {
            "email": "rifarca@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://desafio.idesa.pyaeveapps.com/v2/public"
        }
    ],
    "paths": {
        "/api/user/auth": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Autentifica los usuarios ",
                "operationId": "0cfa3b01ffbf09b7c0fb963769631e17",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "string": {
                                "summary": "email del usuario",
                                "value": "email@example.com"
                            }
                        }
                    },
                    {
                        "name": "password",
                        "in": "path",
                        "description": "password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "string": {
                                "summary": "password del usuario",
                                "value": "123456"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Si autentica devuelve un JSON con el ACCESS TOKEN .",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Devuelve un Objeto JSON con los DATOS del Usuario y su Access Tokebn valido por 72h",
                                        "value": {
                                            "status": "ok",
                                            "message": "Auteticacion de Usuario validado con Exito",
                                            "data": {
                                                "total": 1,
                                                "detail": {
                                                    "id": 5,
                                                    "nombres": "Martin Martinez",
                                                    "correo": "martin.martinez@idesa.com.py",
                                                    "access_token": "4|gBzzeriN5WIoqPoheZqKCcX8ZFeyETx6O8i9CirB",
                                                    "type_token": "Bearer",
                                                    "value_token": "Bearer 4|gBzzeriN5WIoqPoheZqKCcX8ZFeyETx6O8i9CirB",
                                                    "expiration_token": "2800"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Devuelve un JSON con el error .",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Devuelve un Objeto JSON com el mensaje de error",
                                        "value": {
                                            "status": "Unauthorized",
                                            "message": "No tiene autorizacion para consumir la API"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "AuthBearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "scheme": "https",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "AuthBearer": []
        }
    ]
}